fixed index confusion. Output each layer.
When I extended to 3D I got the intuition about index order mixed up. It should be (Layer, Row, Column) but in the code it is (Row, Column, Layer) and all the names are wrong. What's worse is that layerRange and showStateEPS already use the correct layer.
So variable names are a mess, but it seems to work on trivial samples.